Bring back isPluginRequired
